+++ src/pvenv/subcommands/outvenv.py
@@ -14,7 +14,11 @@ class Command(BaseCommand):
 
         print(f"unset {self._prefix}")
         prefix_size = len(self._prefix) + 1
+        unset_prefix_size = len(self._prefix) + len("_unset") + 1
         for key, value in os.environ.items():
             if key.startswith(f"{self._prefix}_"):
-                print(f"export {key[prefix_size:]}={value}")
+                if key.startswith(f"{self._prefix}_unset_"):
+                    print(f"unset {key[unset_prefix_size:]}")
+                else:
+                    print(f"export {key[prefix_size:]}={value}")
                 print(f"unset {key}")